requirements.txt
fastapi
uvicorn
python-chess
requests
python-multipart
python-dotenv
pywebview
